feat: Creating ranking and implementing ranking logic

+++ backend/java/ava_senai/src/main/java/senai/com/ava_senai/domain/ranking/TimeConsumedScore.java
@@ -2,22 +2,22 @@ package senai.com.ava_senai.domain.ranking;
 
 public enum TimeConsumedScore {
 
-    GREATER_THAN_98(98, 0),
-    GREATER_THAN_90(90, 5),
-    GREATER_THAN_75(75, 10),
-    GREATER_THAN_50(50, 15),
-    GREATER_THAN_25(25, 20),
-    LESS_OR_EQUAL_25(0, 25);
+    GREATER_THAN_98(98, 0d),
+    GREATER_THAN_90(90, 5d),
+    GREATER_THAN_75(75, 10d),
+    GREATER_THAN_50(50, 15d),
+    GREATER_THAN_25(25, 20d),
+    LESS_OR_EQUAL_25(0, 25d);
 
     private final double threshold;
-    private final int points;
+    private final double points;
 
-    TimeConsumedScore(double threshold, int points) {
+    TimeConsumedScore(double threshold, double points) {
         this.threshold = threshold;
         this.points = points;
     }
 
-    public static int valueOf(double mediumPercentTimeConsumed) {
+    public static Double valueOf(double mediumPercentTimeConsumed) {
         // Ordem importa: do maior para o menor
         for (TimeConsumedScore score : values()) {
             if (mediumPercentTimeConsumed > score.threshold) {
